Started working on the Updater function.
This function will serve as a updater for each entity in the system.
the operation will call The Open Muni API and The OSM API and
find out if any changes were made by either of them. and update
this DB and Server.

Deleted some unnecessary JSONP files.

eset HEAD <file>..." to unstage)